+++ AElf.Node/Node/IP2P.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace AElf.Kernel.Node
+{
+    public interface IP2P
+    {
+        Task ProcessLoop();
+    }
+}
